Admin panel is revised
