+++ Frontend/tursu/src/ProductList.js
@@ -2,7 +2,9 @@ import React from 'react';
 import './HomePage.css'
 import { makeStyles } from '@material-ui/core/styles';
 import GridList from '@material-ui/core/GridList';
-import ProductBox, {ProductBoxHorizontal} from "./ProductBox";
+import ProductBox from "./ProductBox";
+import NonverifiedProductBox  from "./NonverifiedProductBox";
+import ProductBoxHorizontal from "./ProductBox";
 import Grid from "@material-ui/core/Grid";
 
 
@@ -21,7 +23,7 @@ const useStyles = makeStyles((theme) => ({
 
 export default function ProductListVertical(props) {
     const classes = useStyles();
-    
+
     return (
         <Grid item container className={classes.root} spacing={6}>
             {props.products.map((product) => (
@@ -45,3 +47,6 @@ export function ProductListHorizontal(props) {
         </Grid>
     );
 }
+
+
+
